Move 2D and 3D particle editors to the 2D and 3D folders
