+++ editor/register_editor_types.cpp
@@ -77,6 +77,7 @@
 #include "editor/scene/2d/camera_2d_editor_plugin.h"
 #include "editor/scene/2d/light_occluder_2d_editor_plugin.h"
 #include "editor/scene/2d/line_2d_editor_plugin.h"
+#include "editor/scene/2d/particles_2d_editor_plugin.h"
 #include "editor/scene/2d/path_2d_editor_plugin.h"
 #include "editor/scene/2d/physics/cast_2d_editor_plugin.h"
 #include "editor/scene/2d/physics/collision_polygon_2d_editor_plugin.h"
@@ -93,6 +94,7 @@
 #include "editor/scene/3d/multimesh_editor_plugin.h"
 #include "editor/scene/3d/node_3d_editor_gizmos.h"
 #include "editor/scene/3d/occluder_instance_3d_editor_plugin.h"
+#include "editor/scene/3d/particles_3d_editor_plugin.h"
 #include "editor/scene/3d/path_3d_editor_plugin.h"
 #include "editor/scene/3d/physics/physical_bone_3d_editor_plugin.h"
 #include "editor/scene/3d/polygon_3d_editor_plugin.h"
@@ -108,7 +110,6 @@
 #include "editor/scene/gui/theme_editor_plugin.h"
 #include "editor/scene/material_editor_plugin.h"
 #include "editor/scene/packed_scene_editor_plugin.h"
-#include "editor/scene/particles_editor_plugin.h"
 #include "editor/scene/resource_preloader_editor_plugin.h"
 #include "editor/scene/sprite_frames_editor_plugin.h"
 #include "editor/scene/texture/bit_map_editor_plugin.h"
